rm unused mods
