+++ query/src/catalogs/catalog.rs
@@ -21,10 +21,10 @@ use common_planners::CreateDatabasePlan;
 use common_planners::DropDatabasePlan;
 
 use crate::catalogs::Database;
-use crate::catalogs::DatabaseEngine;
 use crate::catalogs::TableFunctionMeta;
 use crate::catalogs::TableMeta;
-use crate::datasources::engines::database_factory_registry::EngineDescription;
+use crate::datasources::database_engine::DatabaseEngine;
+use crate::datasources::database_engine_registry::EngineDescription;
 
 /// Catalog is the global view of all the databases of the user.
 /// The global view has many engine type: Local-Database(engine=Local), Remote-Database(engine=Remote)
