+++ query/src/datasources/database/mod.rs
@@ -18,7 +18,5 @@ mod tests;
 
 pub(crate) mod default;
 pub(crate) mod example;
-pub(crate) mod local;
-pub(crate) mod prelude_dbs;
-pub(crate) mod remote;
+pub(crate) mod prelude;
 pub(crate) mod system;
