+++ query/src/datasources/mod.rs
@@ -17,5 +17,9 @@ mod index;
 pub(crate) mod common;
 pub(crate) mod dal;
 pub(crate) mod database;
-pub(crate) mod engines;
+pub(crate) mod database_engine;
+pub(crate) mod database_engine_registry;
+pub(crate) mod store_client;
 pub(crate) mod table;
+pub(crate) mod table_engine;
+pub(crate) mod table_engine_registry;
