+++ query/src/datasources/store_client/store_client_provider.rs
